bootrom: Prevent hart 0 from processing to the Security Monitor.
This is a hack for the hifive unleashed board, because at a later
stage the e51 monitor core (hart 0) is being marked as disabled, thus
causing setup_smp to printk a "CPU with hartid=0 is not available"

The alternative would be to change the _relocate_lottery@fw_base.S
value from 0 to -1 so that hart 0 become unlikely to be the boot
core. This alternative also works.